requirements.txt
keyboard
matplotlib
numpy
scipy
ipykernel
jupyter
notebook==6.1.5
pillow
cmasher
jax
optax
flax
click
scikit-learn
zmq
pfilter @ https://github.com/johnhw/pfilter/archive/refs/tags/0.2.5.zip